Shuffling staff members
Shuffling on build time to ensure the statically generated page is the
same as the rendered one, otherwise react throws an error in the console
and the user can see the page flickering. This means the list will
change on every deploy, but every user will see the same order.